requirements.txt
SQLAlchemy==1.4.36
aiosqlite==0.17.0
python-decouple==3.6
python-telegram-bot==20.0a0
dostoevsky==0.6.0
vaderSentiment==3.3.2
meduza==20.7.2
translators==5.2.1
requests==2.27.1
beautifulsoup4==4.11.1